added "vite.js" library to the program, and solved all the problems about import/export
