+++ gcode/index.html
@@ -2,26 +2,27 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>gcode old</title>
+    <title>Vite App</title>
   </head>
   <body>
-    hello world
+    <div id="app">hello world</div>
+
     <script type="module">
       let canActivateTests = {
         console: true,
       };
 
       if (canActivateTests.console == true) {
-        import("./APP/TEST/test.js").then((myFunction) => {
+        import("./src/APP/TEST/test.js").then((myFunction) => {
           myFunction.default();
         });
       } else {
         console.log("test console are desactivated");
       }
     </script>
-
-    <script></script>
+    
+    <script type="module" src="/main.js"></script>
   </body>
 </html>
