+++ gcode/src/APP/TEST/GcodeAPI_main/GcodeAPI.test.js
@@ -1,4 +1,4 @@
-import GcodeAPI from './../../../API/GcodeAPI.js';
+import GcodeAPI from './../../../API/GcodeApi/GcodeAPI.js';
 
 export default function GcodeApiTest() {
   console.warn("🧪 TEST STARTED: GCODE_API_MAIN 🏃‍♂️")
