+++ gcode/src/APP/TEST/Gcommands/G/G0.test.js
@@ -1,12 +1,13 @@
-import G1 from "../../../../API/Gcommands/G/G1.js";
+import G0 from "../../../../API/GcodeApi/Gcommands/G/G0.js";
 
-export default function G1test() {
-  console.warn("🧪 TEST STARTED: G1 🏃‍♂️")
+export default function G0test() {
+  console.warn("🧪 TEST STARTED: G0 🏃‍♂️")
+
   // -1
-  console.log(new G1({}));
+  console.log(new G0({}));
   // 0
   console.log(
-    new G1({
+    new G0({
       x: 1,
       y: 2,
     }).getCode(),
@@ -14,7 +15,7 @@ export default function G1test() {
 
   // 1
   console.log(
-    new G1({
+    new G0({
       x: 3,
       y: 4,
     }).getCode(),
@@ -22,26 +23,26 @@ export default function G1test() {
 
   // 2
   console.log(
-    new G1({
+    new G0({
       x: 5,
     }).getCode(),
   );
 
   // 3
   console.log(
-    new G1({
+    new G0({
       z: 6,
     }).getCode(),
   );
 
   // 4
   console.log(
-    new G1({
+    new G0({
       x: 7,
       y: 8,
       z: 9,
     }).getCode(),
   );
-
-  console.warn("🧪 TEST FINISHED: G1 🚩")
+  
+  console.warn("🧪 TEST FINISHED: G0 🚩")
 }
